Migrate tests to TS and minor fixes.
