+++ app/react/Templates/components/specs/SimilarProperty.spec.tsx
@@ -1,14 +1,20 @@
+/**
+ * @jest-environment jsdom
+ */
 import React from 'react';
 import { shallow } from 'enzyme';
 import { SimilarProperty } from 'app/Templates/components/SimilarProperty';
+import { SimilarPropertiesProps } from '../SimilarProperty';
 
 describe('SimilarProperty', () => {
-  let component;
-  function render(props) {
-    component = shallow(<SimilarProperty {...props} />);
+  let component: any;
+
+  function render(props: SimilarPropertiesProps) {
+    component = shallow<SimilarProperty>(<SimilarProperty {...props} />);
   }
+
   it('should not alert about a conflict if there are no conflicts', () => {
-    const props = {
+    const props: SimilarPropertiesProps = {
       templateProperty: {
         template: 'template 1',
         type: 'relationship',
@@ -24,14 +30,14 @@ describe('SimilarProperty', () => {
     expect(typeCell.length).toBe(0);
     const cells = component.find('td');
     expect(cells.get(0).props.children[2]).toBe('template 1');
-    expect(cells.get(1).props.children[2]).toBe(' relationship');
+    expect(cells.get(1).props.children[2]).toBe('Relationship');
     expect(cells.get(2).props.children[2]).toBe('entity 1');
     const warningIcon = component.find({ icon: 'exclamation-triangle' });
     expect(warningIcon.length).toBe(0);
   });
 
   it('should alert about a conflict  with type', () => {
-    const props = {
+    const props: SimilarPropertiesProps = {
       templateProperty: {
         template: 'template 1',
         type: 'text',
@@ -43,13 +49,13 @@ describe('SimilarProperty', () => {
     render(props);
     const typeCell = component.find('.conflict');
     expect(typeCell.props().title).not.toBeUndefined();
-    expect(typeCell.get(0).props.children[2]).toBe(' text');
+    expect(typeCell.get(0).props.children[2]).toBe('Text');
     const warningIcon = typeCell.find({ icon: 'exclamation-triangle' });
     expect(warningIcon.get(0)).not.toBeUndefined();
   });
 
   it('should alert about a conflict  with relationType', () => {
-    const props = {
+    const props: SimilarPropertiesProps = {
       templateProperty: {
         template: 'template 1',
         type: 'text',
@@ -62,14 +68,14 @@ describe('SimilarProperty', () => {
     render(props);
     const typeCell = component.find('.conflict');
     expect(typeCell.props().title).not.toBeUndefined();
-    expect(typeCell.get(0).props.children[2]).toBe(' text');
+    expect(typeCell.get(0).props.children[2]).toBe('Text');
     expect(typeCell.get(0).props.children[3]).toBe(' (related)');
     const warningIcon = typeCell.find({ icon: 'exclamation-triangle' });
     expect(warningIcon.get(0)).not.toBeUndefined();
   });
 
   it('should alert about a conflict with content of property', () => {
-    const props = {
+    const props: SimilarPropertiesProps = {
       templateProperty: {
         template: 'template 1',
         type: 'select',
